+++ aquadoggo/src/graphql/client/mod.rs
@@ -1,9 +1,14 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
+mod dynamic_query;
 mod mutation;
 mod query;
 mod response;
+mod static_query;
+mod types;
 
+pub use dynamic_query::DynamicQuery;
 pub use mutation::ClientMutationRoot;
 pub use query::ClientRoot;
 pub use response::NextEntryArguments;
+pub use static_query::StaticQuery;
