--- aquadoggo/src/graphql/client/query/mod.rs
@@ -1,10 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
-mod dynamic_query;
-mod graphql_types;
-mod root;
-mod static_query;
-
-pub use dynamic_query::DynamicQuery;
-pub use root::ClientRoot;
-pub use static_query::StaticQuery;
